refactor(tx-builder): extract common pointers code
